+++ src/components/Modal/Modal.tsx
@@ -1,34 +1,19 @@
 import React from "react";
-import {
-  CookieCategoryDefinition,
-  TailorLabels,
-  TailorColors,
-  TailorCookiesDetails,
-} from "../../types";
+import { Label, TailorColors, TailorCookiesDetails } from "../../types";
 import "../../css/out/rct_style.css";
 import { OutlineButton } from "../Buttons";
 import { ModalHeader } from "./ModalHeader";
 import { ModalBody } from "./ModalBody";
 
 interface ModalProps {
-  labels: TailorLabels;
+  labels: Label;
   colors: TailorColors;
   showModal: boolean;
-  categories: CookieCategoryDefinition[];
-  updateCategories: (categories: CookieCategoryDefinition[]) => void;
   cookies: TailorCookiesDetails;
   toggleModal: () => void;
 }
 
-export const Modal = ({
-  labels,
-  showModal,
-  colors,
-  toggleModal,
-  categories,
-  updateCategories,
-  cookies,
-}: ModalProps) => {
+export const Modal = ({ labels, showModal, colors, toggleModal, cookies }: ModalProps) => {
   return (
     <div
       className={`rct-modal ${
@@ -55,13 +40,7 @@ export const Modal = ({
         <div className="modal-content rct-text-left">
           <ModalHeader toggleModal={toggleModal} labels={labels} colors={colors} />
 
-          <ModalBody
-            colors={colors}
-            labels={labels}
-            categories={categories}
-            updateCategories={updateCategories}
-            cookies={cookies}
-          />
+          <ModalBody colors={colors} labels={labels} cookies={cookies} />
 
           <div className={"rct-flex rct-justify-end rct-py-2 rct-px-2"}>
             <OutlineButton