+++ src/constants/defaultCookies.ts
@@ -1,13 +1,21 @@
-import { TailorCookiesDetails } from "../types";
+import { CookieCategory, TailorCookiesDetails } from "../types";
 
 export const defaultCookies: TailorCookiesDetails = {
-  mandatory: [
+  categories: [
+    CookieCategory.MANDATORY,
+    CookieCategory.MARKETING,
+    CookieCategory.PREFERENCES,
+    CookieCategory.STATISTICS,
+    CookieCategory.UNCLASSIFIED,
+  ],
+  data: [
     {
       domain: "tpn-labs.com",
       title: "rct_cookie_consent",
       description: "This is a mandatory cookie that stores the user's cookie consent preferences.",
       expiration: "1 year",
       type: "HTTP",
+      category: CookieCategory.MANDATORY,
     },
     {
       domain: "tpn-labs.com",
@@ -15,9 +23,8 @@ export const defaultCookies: TailorCookiesDetails = {
       description: "This cookie is used to check if the user has accepted the cookie consent.",
       expiration: "Session",
       type: "HTTP",
+      category: CookieCategory.MANDATORY,
     },
-  ],
-  marketing: [
     {
       domain: "Google",
       title: "IDE",
@@ -27,6 +34,7 @@ export const defaultCookies: TailorCookiesDetails = {
         "present targeted ads to the user.",
       expiration: "1 year",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -34,6 +42,7 @@ export const defaultCookies: TailorCookiesDetails = {
       description: "Used to check if the user's browser supports cookies.",
       expiration: "Session",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -42,6 +51,7 @@ export const defaultCookies: TailorCookiesDetails = {
         "Registers a unique ID that identifies a returning user's device. The ID is used for targeted ads.",
       expiration: "6 months",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -50,6 +60,7 @@ export const defaultCookies: TailorCookiesDetails = {
         "These cookies are used statistics and track conversion rates and Google ad personalisation.",
       expiration: "1 month",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -58,6 +69,7 @@ export const defaultCookies: TailorCookiesDetails = {
         "These cookies are used to collect anonymous website statistics and track conversion rates.",
       expiration: "1 year",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -66,6 +78,7 @@ export const defaultCookies: TailorCookiesDetails = {
         "Google cookie. Airship uses this cookie to track the user's consent preferences.",
       expiration: "20 years",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -74,6 +87,7 @@ export const defaultCookies: TailorCookiesDetails = {
         "Google cookie. By using this cookie, Google is able to make the ads more attractive to users and more ",
       expiration: "1 day",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -82,6 +96,7 @@ export const defaultCookies: TailorCookiesDetails = {
         "Google cookie. Google uses this cookie to track the user's consent preferences.",
       expiration: "2 years",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -89,6 +104,7 @@ export const defaultCookies: TailorCookiesDetails = {
       description: "Google cookie. Protects user data from unauthorized access.",
       expiration: "2 years",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -97,6 +113,7 @@ export const defaultCookies: TailorCookiesDetails = {
         "Google cookie. Security cookie to protect a user's data from unauthorized access.",
       expiration: "2 years",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "Google",
@@ -105,6 +122,7 @@ export const defaultCookies: TailorCookiesDetails = {
         "Google cookie. Cookie is used to protect the user data from unauthorized access.",
       expiration: "3 months",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "tpn-labs.com",
@@ -112,6 +130,7 @@ export const defaultCookies: TailorCookiesDetails = {
       description: "This is a marketing cookie.",
       expiration: "Session",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
     {
       domain: "tpn-labs.com",
@@ -119,15 +138,15 @@ export const defaultCookies: TailorCookiesDetails = {
       description: "This is a marketing cookie.",
       expiration: "Session",
       type: "HTTP",
+      category: CookieCategory.MARKETING,
     },
-  ],
-  preferences: [
     {
       domain: "tpn-labs.com",
       title: "Preferences Cookie",
       description: "This is a preferences cookie. 1",
       expiration: "1 year",
       type: "preferences",
+      category: CookieCategory.PREFERENCES,
     },
     {
       domain: "tpn-labs.com",
@@ -135,13 +154,15 @@ export const defaultCookies: TailorCookiesDetails = {
       description: "This is a preferences cookie. 2",
       expiration: "Session",
       type: "pixel",
+      category: CookieCategory.PREFERENCES,
     },
     {
-      domain: "tpn-labs.com",
+      domain: "tpn-labs.com4",
       title: "Preferences Cookie 3",
       description: "This is a preferences cookie. 3",
       expiration: "Session",
       type: "pixel",
+      category: CookieCategory.PREFERENCES,
     },
     {
       domain: "tpn-labs.com4",
@@ -149,24 +170,23 @@ export const defaultCookies: TailorCookiesDetails = {
       description: "This is a preferences cookie. 4",
       expiration: "Session",
       type: "pixel",
+      category: CookieCategory.PREFERENCES,
     },
-  ],
-  statistics: [
     {
       domain: "tpn-labs.com5",
       title: "Statistics Cookie",
       description: "This is a statistics cookie. 5",
       expiration: "1 year",
       type: "statistics",
+      category: CookieCategory.STATISTICS,
     },
-  ],
-  unclassified: [
     {
       domain: "tpn-labs.com",
       title: "Unclassified Cookie",
       description: "This is an unclassified cookie. 6",
       expiration: "1 year",
       type: "unclassified",
+      category: CookieCategory.UNCLASSIFIED,
     },
   ],
 };