+++ stories/defaults/template.tsx
@@ -1,12 +1,21 @@
-import React from 'react';
+import React from "react";
 // @ts-ignore
 import { Intro } from "./intro";
 import { Story } from "@storybook/react";
-import CookieTailor from "../../src";
+import CookieTailor, { TailorResponse } from "../../src";
 
-export const DefaultTemplate: Story<any> = () => (
-  <>
-    <Intro />
-    <CookieTailor debug={true}/>
-  </>
-);
+export const DefaultTemplate: Story<any> = () => {
+  const acceptFunction = (response: TailorResponse) => {
+    console.log("Accept function called");
+    console.log(response.cookieId);
+    console.log(response.cookieCreation);
+    console.log(response.categories);
+  };
+
+  return (
+    <>
+      <Intro />
+      <CookieTailor debug={true} onAccept={acceptFunction} />
+    </>
+  );
+};