+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 > The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [[0.8.0](https://github.com/TPN-Labs/react-cookie-tailor/releases/tag/0.8.0)]
+## [[1.0.0](https://github.com/TPN-Labs/react-cookie-tailor/releases/tag/1.0.0)]
 
 This is the "official" release of the library.
 
@@ -14,12 +14,14 @@ This is the "official" release of the library.
 - A random ID for the consent tracking. This is useful for processing the cookie request
 in the backend. The ID is stored as a cookie with the name `rct_cookie_id`.
 - Settings screen for controlling the cookies
+- Implemented `useCategoryContext` hook for controlling the categories
 - NPM badges in the README.md
 
 ### Updated
 
 - Every cookie created now has the `rct_` prefix
 - Created a prop called `colors` to control the colors of the library
+- Added `response` property on the `onAccept` and `onDecline` callbacks
 
 ### Removed